visits_tags.py edited online with Bitbucket
Currently, the get_visits template tag only allows one to check the count of visits against either an object (e.g request) and a model instance, but I have found a valid use case where it's necessary to count all visits matching a given url regex (or just a url string). I've implemented the regex lookup as the default when the lookup parameter is a string type.